fixes for removing

+++ ghc.py
@@ -52,10 +52,12 @@ if args.l_runs is not None:
 GHC.classifyChannels()
 
 if args.dump != None:
-  os.remove(args.dump)
+  if os.path.exists(args.dump):
+    os.remove(args.dump)
   GHC.Export(args.dump)
 if args.dumpsql != None:
-  os.remove(args.dumpsql)
+  if os.path.exists(args.dumpsql):
+    os.remove(args.dumpsql)
   Data.DumpSQL(GHC.dbh, args.dumpsql)
 
 print "="*80
